services/ci-monitor/requirements.txt
PyGithub>=2.1.1
anthropic>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
